src/requirements.txt
fastapi
uvicorn
jinja2
pyaml
cryptography
sqlalchemy
databases
psycopg2-binary
asyncpg
python-multipart
sqlmodel